requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.75.0
pymupdf==1.25.3
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
aiofiles==24.1.0
